CDMS-1342 moved ModelExtensions into the extensions folder
